requirements.txt
alembic==1.18.4
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.3
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.62.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.10.8
mysql-connector-python==9.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
